+++ examples/README.rst
@@ -1,0 +1,20 @@
+Welcome to the FEDOT examples.
+We provide it to help new users and developers use FEDOT.
+Look at the structure provide below:
+
+simple - for new users
+
+advanced - for advanced users
+
+- simple
+   - classification - devoted to solving basic classification tasks via API
+   - regression - devoted to solving basic regression tasks via API
+   - time_series_forecasting - devoted to solving basic time series forecasting tasks via API
+   - interpretable - shows how pipelines can be interpreted
+- advanced
+   - automl - shows how other autoML solutions can be compatible with FEDOT
+   - decompose - examples devoted to decomposition
+   - fedot_based_solution - shows how FEDOT can be used with non-datascience tasks
+   - remote_execution - shows how FEDOT can be run remotely
+   - sensitivity_analysis - analysis of pipeline structure
+   - time_series_forecasting - advanced examples of time series forecasting
