+++ examples/advanced/automl/pipeline_from_automl.py
