+++ examples/advanced/fedot_profiler_example.py
