+++ examples/advanced/multi_modal_pipeline.py
