+++ examples/simple/pipeline_and_history_visualisation.py
