+++ test/unit/pipelines/test_multi_modal.py
@@ -1,6 +1,6 @@
 import os
 
-from examples.multi_modal_pipeline import (prepare_multi_modal_data)
+from examples.advanced.multi_modal_pipeline import (prepare_multi_modal_data)
 from fedot.core.data.multi_modal import MultiModalData
 from fedot.core.pipelines.node import PrimaryNode, SecondaryNode
 from fedot.core.pipelines.pipeline import Pipeline
